requirements.txt.txt
requests==2.26.0
mysql-connector-python==8.0.25
PyYAML==5.4.1
pandas==1.3.3
